+++ packages/theme_tailor/CHANGELOG.md
@@ -2,7 +2,7 @@
 - Fix name collision when theme class contains `themes` field
 - Fix the bug where the generator could not handle encoders extending generic type encoders
 - Empty theme class will now generate with const constructor
-- Add option to generate constant themes
+- Generate constant themes if possible
 
 # 1.0.3
 - Improve documentation,